+++ src/core/system_impl.h
@@ -63,7 +63,8 @@ public:
     void unregister_mavlink_message_handler(uint16_t msg_id, const void* cookie);
     void unregister_all_mavlink_message_handlers(const void* cookie);
 
-    void register_timeout_handler(std::function<void()> callback, double duration_s, void** cookie);
+    void register_timeout_handler(
+        const std::function<void()>& callback, double duration_s, void** cookie);
     void refresh_timeout_handler(const void* cookie);
     void unregister_timeout_handler(const void* cookie);
 
@@ -244,8 +245,6 @@ private:
     void system_thread();
     void send_heartbeat();
 
-    void process_user_callbacks_thread();
-
     // We use std::pair instead of a std::optional.
     std::pair<MAVLinkCommands::Result, MAVLinkCommands::CommandLong>
     make_command_flight_mode(FlightMode mode, uint8_t component_id);
@@ -288,7 +287,6 @@ private:
     CommandResultCallback _command_result_callback{nullptr};
 
     std::thread* _system_thread{nullptr};
-    std::thread* _process_user_callbacks_thread{nullptr};
     std::atomic<bool> _should_exit{false};
 
     static constexpr double _HEARTBEAT_TIMEOUT_S = 3.0;
@@ -307,27 +305,9 @@ private:
 
     Timesync _timesync;
 
-    TimeoutHandler _timeout_handler;
     CallEveryHandler _call_every_handler;
 
     MAVLinkMissionTransfer _mission_transfer;
-
-    struct UserCallback {
-        UserCallback() {}
-        UserCallback(const std::function<void()>& func_) : func(func_) {}
-        UserCallback(
-            const std::function<void()>& func_,
-            const std::string& filename_,
-            const int linenumber_) :
-            func(func_),
-            filename(filename_),
-            linenumber(linenumber_)
-        {}
-
-        std::function<void()> func{};
-        std::string filename{};
-        int linenumber{};
-    };
 
     std::mutex _plugin_impls_mutex{};
     std::vector<PluginImplBase*> _plugin_impls{};
@@ -342,9 +322,6 @@ private:
     std::function<bool(mavlink_message_t&)> _outgoing_messages_intercept_callback{nullptr};
 
     std::atomic<FlightMode> _flight_mode{FlightMode::Unknown};
-
-    SafeQueue<UserCallback> _user_callback_queue{};
-    bool _callback_debugging{false};
 };
 
 } // namespace mavsdk
